Changes to get PyFerret to compile and link; no warnings about missing function prototypes
git-svn-id: file:///home/users/tmap/svn/repos/ferret/branches/pyferret@14357 fdbf22ae-c210-0410-be80-ca943da6b8f8

+++ fer/common/ferret_shared_buffer.h
@@ -43,8 +43,8 @@ typedef sharedMem *smPtr;
 
 __global smPtr sBuffer;
 
-__global float *memory;
-__global float *ppl_memory;   /* new 10/01 */
+__global double *memory;
+__global double *ppl_memory;   /* new 10/01 */
 
 
 #undef __global
